Make tests work
