+++ mula/.ci/docker-compose.yml
@@ -10,6 +10,9 @@ services:
       - .:/app/scheduler
     env_file:
       - .ci/.env.test
+    depends_on:
+      ci_postgres:
+        condition: service_healthy
 
   ci_scheduler_integration:
     build:
