+++ mula/tests/mocks/__init__.py
@@ -1,0 +1,1 @@
+from .services import MockKatalogusService
